requirements.txt
PyYAML
tqdm
numpy
unidecode
transformers=3.2.0
